Refactored the file structure
Moved the files into seperate folders for routes
