+++ arduino_backend/main.py
@@ -1,7 +1,8 @@
 from fastapi import FastAPI
 from starlette.middleware.cors import CORSMiddleware
 
-from arduino_backend.routers import user, auth
+from arduino_backend.user import router as user_router
+from arduino_backend.auth import router as auth_router
 
 app = FastAPI()
 app.add_middleware(CORSMiddleware,
@@ -11,5 +12,5 @@ app.add_middleware(CORSMiddleware,
                    allow_headers=["*"],
                    )
 
-app.include_router(user.router, prefix="/user")
-app.include_router(auth.router, prefix="/auth")
+app.include_router(user_router.router, prefix="/user")
+app.include_router(auth_router.router, prefix="/auth")
